requirements.txt
flask==2.0.2
bs4==0.0.1
request==2.27.1